% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsAtlasBulk-package.R
\docType{package}
\name{OmicsAtlasBulk-package}
\alias{OmicsAtlasBulk}
\alias{OmicsAtlasBulk-package}
\title{OmicsAtlasBulk: Bulk RNA-seq Pipeline for the OmicsAtlas Project}
\description{
Alignment orchestration (STAR), quantification (featureCounts via Rsubread), differential expression (DESeq2 and edgeR run in parallel with an explicit method-comparison step), functional enrichment (clusterProfiler GO/KEGG/GSEA), and BayesPrism deconvolution against the OmicsAtlas scRNA-seq signature, for the bulk RNA-seq arm of the OmicsAtlas project (see the parent repository for full context).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NosakhareOsaro/omicsatlas}
  \item \url{https://nosakhareosaro.github.io/omicsatlas/r_package/OmicsAtlasBulk/}
  \item Report bugs at \url{https://github.com/NosakhareOsaro/omicsatlas/issues}
}

}
\author{
\strong{Maintainer}: Nosakhare Odionfo Osaro \email{osaronosakhare@gmail.com}

Authors:
\itemize{
  \item Nosakhare Odionfo Osaro \email{osaronosakhare@gmail.com}
}

}
\keyword{internal}
